+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Added missing resource_dfn_clone() arguments to Resource.clone()
+
 ## [1.26.1] - 2025-08-26
 
 ### Added